Initial Next.js frontend (#176)
Add `web-next` as a prototype frontend to replace the Remix app.

This is an experiment in adopting an RSC-native approach, and backing away from our progressive SPA goals. If we move forward we'll rip out the current `apps/web` implementation and add a Expo-based Native app to bring in the more traditional experience we're after.

If we do not move forward this does contain several changes to pages such as country details and entity lists that would need backported.