Initial scaffolding for an Expo-based mobile app (#192)
This kicks off our work on building a native mobile app. Its barebones at the moment, but has a lot of scaffolding up and running.

This also upgrades us to pnpm 9.5.0 catalogs in a variety of places, downgrades Next to 14 and React to 18 (RN and Expo are not yet compatible with React 19), and fixes a long-standing warning we had with misconfigured Tailwind.

Heavily inspired by create-t3-turbo's scaffolding: https://github.com/t3-oss/create-t3-turbo

Stil some work to do here, and my goal is to remove the hoisting as soon as I can, and we may overhaul the core of the mobile repo.