Expose registration flows (#197)
Add a register and email verification flow, as well as the related supporting components. Refactors a variety of things, including changing the overall design of forms.

Note this also fixes jsx-email, which was broken in dev/prod (passing in tests). There was an issue with react-jsx in typescript which I couldn't figure out, so instead we opted to switch to react for the jsx option. The specific error was React is not defined when executing the code in the server application.

This is missing a reset password flow, which will be a followup.