Various releases behaviors (#260)
This adds a large amount of the behaviors needed to support releases
(previously known as editions).

It includes release selection on the tasting form (might tweak this
later), various changes to endpoints to add the optional support of
releases, and a variety of related cleanup.

There's still more work to do here, but this PR is already getting
larger than it was intended.

Follows, at minimum, include:

- Add 'addRelease' flow (requires bottle input).
- Add release selection (or creation) to 'addBottle' flow.
- Add dedicated release subpage of bottles (/bottles/ID/releases/ID)
- Expand search to include precision release matches (see below)

And once the above is done, and any related work, most importantly we'll
need to migrate existing vintages/limited releases into the new table.

Refs GH-246