Add bottle series relations (#269)
A new parent relation for Bottles which will define a grouping mechanism
for expressions in the same series.

This is useful for bottlings like High West's "A Midwinter Night's
Dram". There is an ~annual release of this bottle and we want to ensure
its easy to discover (and collect!) each release.

Eventually we're likely to remove BottleRelease and stick with the
original Bottle design for releases, utilizing the new Series grouping
mechanism to provide a cleaner path to simplifying search.

Refs GH-246